Added cmake function to create tests
The main purpose for this is to automatically set the PROCESSORS
and PROCESSOR_AFFINITY properties of the test, to allow running
'cmake -jN' correctly. That means cmake does not assume all tests
use 1 processor, so it can better schedule how many to run concurrently.

The function also allows to make parsing our CMakeLists.txt files
a bit more easy.